# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Digital City Delivery App API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ServiceCategory =
  (typeof ServiceCategory)[keyof typeof ServiceCategory];

export const ServiceCategory = {
  restaurant: "restaurant",
  pharmacy: "pharmacy",
  lawyer: "lawyer",
  grocery: "grocery",
  mechanic: "mechanic",
  doctor: "doctor",
} as const;

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  pending: "pending",
  confirmed: "confirmed",
  in_progress: "in_progress",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export interface ServiceProvider {
  id: number;
  name: string;
  nameAr: string;
  category: ServiceCategory;
  description: string;
  descriptionAr: string;
  address: string;
  rating?: number;
  isAvailable: boolean;
  createdAt: string;
}

export interface CreateOrderRequest {
  customerName: string;
  customerAddress: string;
  notes?: string;
  serviceProviderId: number;
  serviceType: ServiceCategory;
}

export interface UpdateOrderStatusRequest {
  status: OrderStatus;
}

export interface Order {
  id: number;
  customerName: string;
  customerAddress: string;
  notes?: string;
  serviceType: ServiceCategory;
  status: OrderStatus;
  serviceProviderId: number;
  serviceProviderName: string;
  createdAt: string;
  updatedAt: string;
}

export interface ErrorResponse {
  message: string;
}

export type ListServicesParams = {
  category?: ServiceCategory;
};