# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Digital City Delivery App API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: orders
    description: Order management
  - name: services
    description: Service categories
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List all orders
      description: Returns all orders for the admin dashboard
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place a new order
      description: Customer submits an order to a service provider
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusRequest"
      responses:
        "200":
          description: Order updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /services:
    get:
      operationId: listServices
      tags: [services]
      summary: List all service providers
      parameters:
        - name: category
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/ServiceCategory"
      responses:
        "200":
          description: List of service providers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ServiceProvider"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    ServiceCategory:
      type: string
      enum:
        - restaurant
        - pharmacy
        - lawyer
        - grocery
        - mechanic
        - doctor
    OrderStatus:
      type: string
      enum:
        - pending
        - confirmed
        - in_progress
        - delivered
        - cancelled
    ServiceProvider:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        nameAr:
          type: string
        category:
          $ref: "#/components/schemas/ServiceCategory"
        description:
          type: string
        descriptionAr:
          type: string
        address:
          type: string
        rating:
          type: number
          format: float
        isAvailable:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - nameAr
        - category
        - description
        - descriptionAr
        - address
        - isAvailable
        - createdAt
    CreateOrderRequest:
      type: object
      properties:
        customerName:
          type: string
        customerAddress:
          type: string
        notes:
          type: string
        serviceProviderId:
          type: integer
        serviceType:
          $ref: "#/components/schemas/ServiceCategory"
      required:
        - customerName
        - customerAddress
        - serviceProviderId
        - serviceType
    UpdateOrderStatusRequest:
      type: object
      properties:
        status:
          $ref: "#/components/schemas/OrderStatus"
      required:
        - status
    Order:
      type: object
      properties:
        id:
          type: integer
        customerName:
          type: string
        customerAddress:
          type: string
        notes:
          type: string
        serviceType:
          $ref: "#/components/schemas/ServiceCategory"
        status:
          $ref: "#/components/schemas/OrderStatus"
        serviceProviderId:
          type: integer
        serviceProviderName:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - customerName
        - customerAddress
        - serviceType
        - status
        - serviceProviderId
        - serviceProviderName
        - createdAt
        - updatedAt
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message