# lib/api-zod/src/generated/types/createOrderRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Digital City Delivery App API
 * OpenAPI spec version: 0.1.0
 */
import type { ServiceCategory } from "./serviceCategory";

export interface CreateOrderRequest {
  customerName: string;
  customerAddress: string;
  notes?: string;
  serviceProviderId: number;
  serviceType: ServiceCategory;
}