# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Digital City Delivery App API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderStatus } from "./orderStatus";
import type { ServiceCategory } from "./serviceCategory";

export interface Order {
  id: number;
  customerName: string;
  customerAddress: string;
  notes?: string;
  serviceType: ServiceCategory;
  status: OrderStatus;
  serviceProviderId: number;
  serviceProviderName: string;
  createdAt: Date;
  updatedAt: Date;
}