# lib/api-zod/src/generated/types/serviceCategory.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Digital City Delivery App API
 * OpenAPI spec version: 0.1.0
 */

export type ServiceCategory =
  (typeof ServiceCategory)[keyof typeof ServiceCategory];

export const ServiceCategory = {
  restaurant: "restaurant",
  pharmacy: "pharmacy",
  lawyer: "lawyer",
  grocery: "grocery",
  mechanic: "mechanic",
  doctor: "doctor",
} as const;