# lib/api-zod/src/generated/types/serviceProvider.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Digital City Delivery App API
 * OpenAPI spec version: 0.1.0
 */
import type { ServiceCategory } from "./serviceCategory";

export interface ServiceProvider {
  id: number;
  name: string;
  nameAr: string;
  category: ServiceCategory;
  description: string;
  descriptionAr: string;
  address: string;
  rating?: number;
  isAvailable: boolean;
  createdAt: Date;
}